% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dobeta.R
\name{dobeta}
\alias{dobeta}
\title{Function to determine the MUE estimate based on the beta distribution.}
\usage{
dobeta(n, y)
}
\arguments{
\item{n}{Number of independent trials.}

\item{y}{Number of successes.}
}
\value{
A number with the MUE estimate
}
\description{
This function calculates the median unbiased estimate for a single proportion.
}
