% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mue_confidence_interval.R
\name{mue_confidence_interval}
\alias{mue_confidence_interval}
\title{Generate the exact bootstrap interval for the MUE estimate.}
\usage{
mue_confidence_interval(n1, p1, n2, p2, alpha)
}
\arguments{
\item{n1}{Sample size for group 1.}

\item{p1}{MUE estimate calculated by the dobeta function for group 1.}

\item{n2}{Sample size for group 2.}

\item{p2}{MUE estimate calculated by the dobeta function for group 2.}

\item{alpha}{The significance level for the confidence interval.}
}
\value{
Returns a matrix of the upper and lower limits for the confidence interval.
}
\description{
This function enumerates the exact bootstrap confidence interval for the MUE-based estimate of the relative risk.
}
